feat: Create class for Series
Move util functions relating to series into Series class.
Remove the extra loop for expanding a series from utils.get_list_of_work_ids: since we moved the logic for getting a series's url into the Series class, and Series imports from utils, we were getting a circular import error.
Create a custom get_list_of_work_ids method in User: this method includes the logic for expanding a series. The only paginated list of ids that can contain both work and series ids is the bookmarks list (which belongs to a user and is accessed via the User class), so this is the only place where we need this logic. Still needs to be tidied up, if possible.